Refactor some mouse and window C code to Go

+++ base/bitmap_free_c.h
@@ -2,7 +2,7 @@
 #include <assert.h>
 #include <string.h>
 
-MMBitmapRef createMMBitmap_c(uint8_t *buffer, size_t width, size_t height, 
+MMBitmapRef createMMBitmap_c(uint8_t *buffer, int32_t width, int32_t height, 
 	size_t bytewidth, uint8_t bitsPerPixel, uint8_t bytesPerPixel
 ) {
 	MMBitmapRef bitmap = malloc(sizeof(MMBitmap));
